+++ src/stim/stabilizers/tableau.pybind.cc
@@ -19,9 +19,9 @@
 #include "stim/simulators/tableau_simulator.h"
 #include "stim/stabilizers/conversions.h"
 #include "stim/stabilizers/pauli_string.h"
-#include "stim/stabilizers/pauli_string.pybind.h"
 #include "stim/stabilizers/tableau.h"
 #include "stim/stabilizers/tableau_iter.h"
+#include "stim/util_top/stabilizers_vs_amplitudes.h"
 
 using namespace stim;
 using namespace stim_pybind;
